+++ backend/src/baserow/config/settings/demo.py
@@ -1,4 +1,5 @@
 from .base import *  # noqa: F403, F401
 
 
+DEBUG = True
 EMAIL_BACKEND = 'django.core.mail.backends.console.EmailBackend'