+++ backend/tests/fixtures/__init__.py
@@ -1,6 +1,7 @@
 from faker import Faker
 
 from .user import UserFixtures
+from .user_file import UserFileFixtures
 from .group import GroupFixtures
 from .application import ApplicationFixtures
 from .table import TableFixtures
@@ -9,6 +10,6 @@ from .field import FieldFixtures
 from .token import TokenFixtures
 
 
-class Fixtures(UserFixtures, GroupFixtures, ApplicationFixtures, TableFixtures,
-               ViewFixtures, FieldFixtures, TokenFixtures):
+class Fixtures(UserFixtures, UserFileFixtures, GroupFixtures, ApplicationFixtures,
+               TableFixtures, ViewFixtures, FieldFixtures, TokenFixtures):
     fake = Faker()