+++ web-frontend/modules/database/plugin.js
@@ -9,6 +9,7 @@ import {
   NumberFieldType,
   BooleanFieldType,
   DateFieldType,
+  FileFieldType,
 } from '@baserow/modules/database/fieldTypes'
 import {
   EqualViewFilterType,
@@ -61,6 +62,7 @@ export default ({ store, app }) => {
   app.$registry.register('field', new DateFieldType())
   app.$registry.register('field', new URLFieldType())
   app.$registry.register('field', new EmailFieldType())
+  app.$registry.register('field', new FileFieldType())
   app.$registry.register('importer', new CSVImporterType())
   app.$registry.register('importer', new PasteImporterType())
   app.$registry.register('settings', new APITokenSettingsType())